fix: Improve file structure
